\name{DeterminingElementForThresholdCalculations}
\alias{DeterminingElementForThresholdCalculations}
\title{DeterminingElementForThresholdCalculations}
\usage{
DeterminingElementForThresholdCalculations(DataframeOfDeconvolutedCounts)
}
\description{
This function is used by DetermineGBExtents() and will lead to questions being asked in the console that the user can respond to.  This enables the user to plot indiviudal elements to inform their choice of which element to use to calculate the thresholds for the Gibbs Excess calculations.

DataframeOfDeconvolutedCounts is a dataframe containing Distance and count of each element with distance.  Can be generated using SplitComplexIons() function.
}
\examples{

}
