\name{ThresholdSuitableCheck}
\alias{ThresholdSuitableCheck}
\title{ThresholdSuitableCheck}
\usage{
ThresholdSuitableCheck()
}
\description{
This function is used by DetermineGBExtents() and will lead to questions being asked in the console that the user can respond to.  This enables the user to check that sensible thresholds are calculated for a given p-value, element selection, and start and end of where matrix definitely exists up to.
}
\examples{

}
