\name{SplitComplexIons}
\alias{SplitComplexIons}
\title{SplitComplexIons}
\usage{
SplitComplexIons()
}
\description{
This function breaks complex ions into their consituent parts.  For example, MoC into Mo and C or H2O into 2H and O.  This is useful when trying to plot changes in concentration of a certain element with distance.

}
\examples{

}
