\name{ThresholdSuitableCheck}
\alias{ThresholdSuitableCheck}
\title{ThresholdSuitableCheck}
\usage{
ThresholdSuitableCheck()
}
\description{
This function is used by DetermineGBExtents() and allows the user state if sensible thresholds have been calculated or not.  If not,     DeterminingElementForThresholdCalculations() will be performed again.
}
\examples{

}
